Moves handling of settings to config object
