+++ web/src/app/shared/classes/websocket.ts
@@ -7,14 +7,24 @@ export class WebSocketMessage {
     public static CreateAuthenticationMessage(): WebSocketMessage {
         var response = new WebSocketMessage();
         
-        response.event = "auth";
+        response.event = WebSocketMessageType.Auth;
         response.data = { "username": window.sessionStorage["username"], "password": window.sessionStorage["password"] };
 
         return response;
     };
+
+    public static GetSettingsMessage() : WebSocketMessage {
+        var response = new WebSocketMessage();
+
+        response.event = WebSocketMessageType.GetConfig;
+        response.data = null;
+
+        return response;
+    }
 }
 
 export class WebSocketMessageType {
+    public static Auth: string = "auth";
     public static GetConfig: string = "GetConfig";
     public static SaveConfig: string = "SaveConfig";
     public static GetPortfolio: string = "GetPortfolio";
